Updated stream URLs for Fip Radio
Their "pretty" URL is http://direct.fipradio.fr/live/fip-midfi.mp3 but is has two built-in HTTP 302 redirects which the current (aka. old) version of mpd which is in the Raspbian repositories can't handle. So I use wget on the pretty URL to see the final, real URL, now updated in this file.